rename: fetchMovieList API 파일을 movieRequests 네이밍으로 변경
